Add files via upload
python files for CS4250 homework 4, questions 4 and 5.

crawler.py is for question 4 and should strictly follow the provided pseudocode

parser.py is for question 5. it uses pymongo and beautifulsoup to get the HTML from the existing database and parse it for the professor information. the professors' information is then stored in a new collection called "professors" in which each professor has their own document